Introduce CIFChatViewer
This is a big one.
We can now modify the chat. In general, the original chat is overwritten by our own implementation. We do this by using OverrideObject to replace the Create and Delete-functions of CIFChatViewer in the original client.
From there on, our own virtual function table with our own functions is used; therefore we have full control over the whole CIFChatViewer. I implemented almost all chat functions. The receive function is placed in unsorted.cpp. It was originally placed in CNetProcessIn, but we don't have that yet.