Add various environment checks to project generation script
The script was very simplistic and could not deal with errors at all. We're now checking for git and cmake being executable, for Visual Studio being installed, the compiler being present and the compiler being able to compile a simple binary with windows sdk dependencies. This should rule out most of the issues people have when deciding they don't need to read the "Getting started" page.